moved function complex_phase_value above one level to avoid circular import and fixed accordingly other files
