+++ src/sympleq/applications/measurement/aquire_utils.py
@@ -1,8 +1,9 @@
 import numpy as np
+from sympleq import complex_phase_value
 from sympleq.core.paulis import PauliSum
 from sympleq.applications.measurement.allocation import scale_variances
 from sympleq.applications.measurement.covariance_graph import graph
-from sympleq.core.paulis.utils import complex_phase_value, covariance_matrix
+from sympleq.core.paulis.utils import covariance_matrix
 # from sympleq.core.statistic_utils import true_covariance_graph
