+++ src/sympleq/core/paulis/utils.py
@@ -386,20 +386,3 @@ def covariance_matrix(P: PauliObject, psi: np.ndarray, include_weights: bool = F
             for i0 in range(n_paulis)]
     )
     return covariance_matrix
-
-
-def complex_phase_value(phase: int, dimension: int, full: bool = True) -> complex:
-    """
-    Computes the a-th eigenvalue of a pauli with dimension d.
-
-    Args:
-        phase (int): The integer to compute the eigenvalue for.
-        dimension (int): The dimension of the pauli to use.
-
-    Returns:
-        complex: The computed eigenvalue.
-    """
-    if full:
-        return np.exp(2 * np.pi * 1j * phase / (2 * dimension))
-    else:
-        return np.exp(2 * np.pi * 1j * phase / dimension)
